Two-Sample Comparison of Average Hazard

The ah2 function performs two-sample comparisons using the average hazard (AH) as a summary measure of the survival time distribution. Two kinds of between-group contrast metrics, the ratio of AH (RAH) and the difference in AH (DAH), are calculated.

Usage

ah2(time, status, arm, tau=NULL, conf.int=0.95)

Arguments

time

The follow-up time for right censored data.

status

The event indicator, 1=event, and 0=right censored.

arm

The group indicator for comparison. The elements of this vector take either 1 or 0. Normally, 0=control group, 1=active treatment group.

tau

A scalar value to specify a time point for calculating the average hazard. When tau=NULL, the default value (i.e., the maximum time point where the size of risk set for both groups remains at least 10) is used.

conf.int

A confidence coefficient for calculating confidence intervals. The default is conf.int=0.95.

Value

an object of class ah2.

note

the truncation time point used in the analysis

n.obs

the number of observations (total number, number of events by tau, number of censoring by tau, and size of risk set at tau)

ah

the estimated average hazard by arm

rah

the ratio of average hazards (RAH; treatment over control)

dah

the difference of average hazard (DAH; treatment minus control)

Details

The function provides the AH for each of the two groups, the absolute difference and the absolute ratio of AH (DAH and RAH) between the two groups, and the corresponding confidence intervals. It also calculates p-values for the two-sided tests based on the RAH and DAH.

References

Horiguchi M and Uno H. Ratio and difference of average hazard with survival weight: new measures to quantify survival benefit of new therapy. Statistics in Medicine. 2023;1-17. <doi:10.1002/sim.9651>

Author

Hajime Uno, Miki Horiguchi

Examples

#====================================================================
# cm214_pfs: The sample reconstructed data of the CheckMate214 study.
# The code below reproduces the results reported by
# Uno and Horiguchi (StatMed; 2023) in Table 6.
#====================================================================
D      = cm214_pfs
time   = D$time
status = D$status
arm    = D$arm
tau    = 21

a = ah2(time=time, status=status, arm=arm, tau=tau, conf.int=0.95)
print(a, digits=3)
#> 
#> The truncation time: tau = 21 was specified.
#> 
#> Number of observations: 
#>      Total N Event by tau Censor by tau At risk at tau
#> arm0     422          225           163             34
#> arm1     425          219           160             46
#> 
#> 
#> Average Hazard (AH) by arm: 
#>            Est. Lower 0.95 Upper 0.95
#> AH (arm0) 0.066      0.057      0.076
#> AH (arm1) 0.049      0.042      0.057
#> 
#> 
#> Between-group contrast: 
#>                                Est. Lower 0.95 Upper 0.95 P-value
#> Ratio of AH (arm1/arm0)       0.747      0.608      0.917   0.005
#> Difference of AH (arm1-arm0) -0.017     -0.029     -0.005   0.006